pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pomodoro-shell-cli"
version = "0.1.0"
description = "Command-line interface for GNOME Pomodoro"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "dbus-python",
    "PyGObject",
]

[project.scripts]
pomodoro = "pomodoro_shell_cli.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["pomodoro_shell_cli*"]
